[Scorm] cleans evaluation management
