+++ src/plugin/scorm/Resources/config/services.yml
@@ -1,5 +1,4 @@
 imports:
-    - { resource: services/command.yml }
     - { resource: services/controller.yml }
     - { resource: services/finder.yml }
     - { resource: services/listener.yml }
@@ -7,9 +6,4 @@ imports:
     - { resource: services/serializer.yml }
 
 services:
-    _defaults:
-        autowire: false
-        autoconfigure: false
-        public: true
-
     Claroline\ScormBundle\Library\ScormLib: ~
