+++ src/plugin/scorm/Resources/config/services/finder.yml
@@ -1,9 +1,4 @@
 services:
-    _defaults:
-        autowire: false
-        autoconfigure: false
-        public: true
-
     _instanceof:
         Claroline\AppBundle\API\Finder\AbstractFinder:
             tags: [claroline.finder]
