+++ src/plugin/scorm/Resources/config/services/serializer.yml
@@ -1,8 +1,5 @@
 services:
     _defaults:
-        autowire: false
-        autoconfigure: false
-        public: true
         tags: [claroline.serializer]
 
     Claroline\ScormBundle\Serializer\ScormSerializer:
